// src/client/apps/MoneyClicker/src/ImageReferences.ts
export enum MoneyClickerPictureReferences {
  MessageBoxPic = 'assets/images/moneyClicker/messageBox.png',

  RotatingThingPic = 'assets/images/moneyClicker/rotatingThing.png',

  VorPic = 'assets/images/moneyClicker/vor.png',
  SmallBangPic = 'assets/images/moneyClicker/smallBangLogoShow.png',
  UpgradePic = 'assets/images/moneyClicker/upgrade.png',

  VorTop = 'assets/images/moneyClicker/vorLabel.png',
  VorMid = 'assets/images/moneyClicker/vorWindow.png',
  VorBotP = 'assets/images/moneyClicker/vorBottom.png',
  VorBotL = 'assets/images/moneyClicker/vorBottomL.png',
  VorL = 'assets/images/moneyClicker/vorList.png',

  SBTop = 'assets/images/moneyClicker/smallBangLabel.png',
  SlBMid = 'assets/images/moneyClicker/smallBangWindow.png',
  SBBotP = 'assets/images/moneyClicker/smallBangBottom.png',
  SBBotL = 'assets/images/moneyClicker/smallBangBottomL.png',
  SBL = 'assets/images/moneyClicker/smallBangList.png',

  UpTop = 'assets/images/moneyClicker/upgradesLabel.png',
  UpBMid = 'assets/images/moneyClicker/upgradesWindow.png',
  UpBotP = 'assets/images/moneyClicker/upgradesBottom.png',
  UpBotL = 'assets/images/moneyClicker/upgradesBottomL.png',
  UpL = 'assets/images/moneyClicker/upgradesList.png',
  BlackHoleImg = 'assets/images/moneyClicker/blackHole.png',

  Money1 = 'assets/images/moneyClicker/gnar1.png',
  Money5 = 'assets/images/moneyClicker/gnar5.png',
  Money10 = 'assets/images/moneyClicker/gnar10.png',
  Money20 = 'assets/images/moneyClicker/gnar20.png',
  Money50 = 'assets/images/moneyClicker/gnar50.png',
  Money100 = 'assets/images/moneyClicker/gnar100.png',
  Money200 = 'assets/images/moneyClicker/gnar200.png',
  Money500 = 'assets/images/moneyClicker/gnar500.png',
}